requirements.txt
aiogram==3.*
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
pydantic-settings>=2.3
apscheduler>=3.10
asyncpg>=0.29
SQLAlchemy[asyncio]>=2.0
alembic>=1.13
python-dotenv>=1.0
orjson>=3.10
structlog>=24.1
pytest>=8.2
ruff>=0.5
black>=24.8
mypy>=1.10
